Remove two superfluous calls in `.aleo` parser.
The parsing of the two required entries of record types (owner and gates) was
doing a superfluous call to `Sanitize::parse()` (which skips over comments and
whitespace) just before parsing the entry type. But this call was already
preceded by a call to `Sanitize::parse_whitespace()`, which is the appropriate
one in this part of the syntax, consistently with the parsing of other (i.e.
non-required) record entries in the `parse_entry` function.