Disallow comments in a certain place.
In the rest of the parser, comments and whitespace are allowed before each
construct that it typically written in its own line (e.g. instruction, input,
etc.), while whitespace only (not comments) is allowed in the middle of such
constructs (i.e. in the middle of the line, e.g. between two operands). There
are two exceptions to this pattern: one is removed by
https://github.com/AleoHQ/snarkVM/pull/1174; the other is removed by this commit
and PR. This one applies to what goes between the last (if any) operand of a
`finalize` command and the ending semicolon.