Revert Query to original state
This reverts the attempt to use the inclusion query object to pass in
outside data from a WASM context. The reasoning behind this is that it's
difficult to know ahead of time which commitments will be needed to
build the proof. It is likely for full WASM support, a different set of
composing functions will be needed.