Add a test related to field from bits.
This is actually to test the constraint subsystem that checks whether the
integer value of the bits is not above a known constant value, which for this
field operation is the prime minus one, but this test tries that with a smaller
number that has some ones as the least significant bits. This confirms what can
be inferred by looking at the code, namely that the least significant one bits
are skipped, and constraints start only at the first zero bit; furthermore, the
first constraint is slightly different than the subsequent ones.