Use ProvingKey and circuit_hash as mapping keys
In order to avoid passing ProvingKey everywhere (which doesn't only require memory but also changing downstream code to be templated on E: PairingEngine, I split this up into various smaller maps, using a reference to or the actual circuit hash as key (useful for serialization and identification in witness labels), and "other stuff" as values. I suggest at the end we do a pass through that "other stuff" to see what consolidation makes sense, but happy to hear your thoughts on the whole setup.