Remove custom thread pool initialization from wasm parallel contexts
When running multi-threaded wasm a global thread pool is initialized using Rayon's event loop
running on webworker threads prior to any work done on these threads. The execute_with_max_threads
function normally spins up worker threads on a separate thread pool, but since the webworkers
are working on the global thread pool, this is redundant. It may be possible to spin up a separate
thread pool in this operation but it's not gaurunteed to necessarily be more performant in the web
and would be a matter of testing. This change allows us to use the approach in the wasm-bindgen-rayon
crate to bring multi-threaded wasm proving operations to the web now in a way that allows developers
to predictably control the thread counts.