[Aleo ins parser] Minor code and doc fixes.
This is for futures.

Update some doc comments.

The code fix is for parsing whitespace just before commas of future arguments.
The current code allows comments and new lines there, which I don't think we
want (not wrong per se, but anomalous, inconsistent with other parts of the
syntax). So the fix restricts that to just whitespace (no comments) without new
lines. We still allow comments and new lines after the comma, consistently with
other parts of the syntax.