[Aleo ins parser] Fix in struct/array parsing.
This is fairly minor, but the current parser disallows whitespace between a
component of a struct or an array and the comma that separates it from the next
component. For instance, this is currently disallowed:
```
{x: 0u8 , y: 1u8}
```
because of the space between `0u8` and `,`. Regardless of whether it is good or
bad style to put a space there, other parts of the Aleo instruction syntax allow
whitespace just before separators like commas or colons. Thus, this PR allows
that.