Have Keccak and SHA-3 operate on any bits.
As discussed on Slack and agreed with Howard, this commit removes the code that
zero-extends the input bits of Keccak and SHA-3 to be byte-aligned. Consistently
with FIPS 202, these hashes are bit-oriented, and operate on bit sequences of
any lengths.

While typical implementations of these hashes often take byte (not bit)
sequences as inputs, our implementation may be used to hash boolean arrays in
the future.  In this case, we want different boolean arrays to hash to different
values, but the zero-extension to byte would cause collisions in that case.

Our current tests compare our implementation with our console wrapper of an
implementation that operates on byte sequences. So we need to zero-extend to
byte in the tests before calling our implementation, as indicated by the added
comments.

At some point we should probably add other tests with known vectors, for both
byte-aligned and non-byte-aligned bit sequences.