Fix circuit for cast lossy field to group.
This resolves the underconstraining bug that was found in that circuit.

The fix involves the creation of two new circuits, which are more general and
may have other uses:
- A circuit for square roots flagged nondeterministic, which return both square
  roots, in no specified order (hence 'nondeterministic'), if the input is a
  non-zero square, and returns an error flag set if the input is not a square.
- A circuit for turning an x coordinate into the (unique, if it exists) point
  in the subgroup with that x coordinate, with an error flag set if there is
  no such point in the subgroup. This circuit makes use of the previous one
  to solve the curve equation, i.e. to find whether and where x intersects
  the elliptic curve.

The new circuit for cast lossy field to group is realized by using the latter
circuit to attempt to find the point in the subgroup with the given value as x
coordinate. If none exists, we return the generator if the input is 0, otherwise
the result of Elligator-2 is returned, which is always a subgroup point.

The use of the new circuits for flagged operations eliminate the
underconstraining without eliminating desired solutions (which would happen by
enforcing constraints without taking flags into account).