Fix bug in Poseidon internal squeezing function.
The permutation must be done unconditionally at that point of the code, which is
reached if there are more field elements to squeeze than are left in the current
state vector (i.e. `squeeze_index + remaining.len() > RATE` holds). Whether the
remaining elements to squeeze happen to be the same number as the rate or not,
is irrelevant here.